requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
readability-lxml==0.8.1
tqdm==4.66.4

sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1

transformers==4.43.3
torch>=2.1.0
accelerate==0.33.0
